app/backend/pyproject.toml
[project]
name = "ai-docs"
version = "0.0.1"
description = "AI-Docs is an AI-powered chat application that performs RAG (Retrieval-Augmented Generation) on popular technology documentation."
readme = "README.md"
authors = [
    { name = "Lokeswaran Aruljothy", email = "lokeswaranaruljothy@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.13",
    "langchain-core>=0.3.28",
    "langchain-openai>=0.2.14",
    "langchain-postgres>=0.0.12",
    "python-dotenv>=1.0.1",
    "psycopg-binary>=3.2.3",
    "langgraph>=0.2.60",
]

[project.scripts]
graph = "graph:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
